-added SheetMetalRollerBlock with its associated boilerplate. (it has no functionality) -made Multiblock stuff more abstract and common -improved the MultiBlockPreviewRenderer and its mixin (don't really need the mixin could be replace with a singleton)

+++ src/main/java/com/nythicalnorm/voxelspaceprogram/block/rocket_parts/entity/EngineProperties.java
@@ -2,8 +2,8 @@ package com.nythicalnorm.voxelspaceprogram.block.rocket_parts.entity;
 
 public class EngineProperties {
     private final int blockSize;
-    private final double pixelHeight;
-    private final double pixelWidth;
+    private final float pixelHeight;
+    private final float pixelWidth;
 
     public EngineProperties(Builder builder) {
         this.blockSize = builder.blockSize;
@@ -15,18 +15,18 @@ public class EngineProperties {
         return blockSize;
     }
 
-    public double getPixelHeight() {
+    public float getPixelHeight() {
         return pixelHeight;
     }
 
-    public double getPixelWidth() {
+    public float getPixelWidth() {
         return pixelWidth;
     }
 
     public static class Builder {
         private int blockSize;
-        private double pixelHeight;
-        private double pixelWidth;
+        private float pixelHeight;
+        private float pixelWidth;
 
         public Builder BlockSize(int size) {
             this.blockSize = size;
